+++ bits/BUILD.bazel
@@ -11,7 +11,6 @@ rust_library(
     visibility = ["//visibility:public"],
     deps = [
         "//bits/bit",
-        "//num",
     ],
 )
